refactor: integrate munimap_transport into munimap module
