+++ munimap/templates/munimap/app/map.html
@@ -173,7 +173,7 @@
        activated-callback="$parent.$parent.onAreaMeasureActivated">
     </a>
     {% if base_config.get("TIMETABLE_DEFAULT_CTIY") == 'Bielefeld' %}
-    <a href="https://beteiligung.nrw.de/portal/bielefeld/beteiligung/themen/1001101" title="Aufruf des Bielefelder Mängelmelders" 
+    <a href="https://beteiligung.nrw.de/portal/bielefeld/beteiligung/themen/1001101" title="Aufruf des Bielefelder Mängelmelders"
       target="_blank" class="list-group-item">
       Mängelmelder BI
     </a>
@@ -279,7 +279,7 @@
       <span class="glyphicon glyphicon-time"></span>
     </button>
   </div>
-  {% include 'transport/app/timetable.html' %}
+  {% include 'transport/timetable.html' %}
   {% endif %}
 </div>
 
@@ -307,7 +307,7 @@
 
 {% if app_config['components']['timetable'] %}
 <script type="text/ng-template" id="munimap-transport/station-properties.html">
-    {% include 'transport/app/ng-templates/station-properties.html' %}
+    {% include 'transport/ng-templates/station-properties.html' %}
 </script>
 {% endif %}
