+++ munimap/templates/munimap_transport/index.html
@@ -24,11 +24,11 @@
     </script>
 
     {% block sidebar %}
-      {% include 'transport/app/sidebar.html' %}
+      {% include 'transport/sidebar.html' %}
     {% endblock %}
 
     {% block map %}
-      {% include 'transport/app/map.html' %}
+      {% include 'transport/map.html' %}
     {% endblock %}
   </div>
