+++ munimap/templates/munimap_transport/map.html
@@ -183,19 +183,19 @@
   <div class="ol-control map-logo">
     <img src="{$ url_for('static', filename='img/logo_transport.png') $}">
   </div>
-  {% include 'transport/app/timetable.html' %}
+  {% include 'transport/timetable.html' %}
 </div>
 
 <script type="text/ng-template" id="munimap-transport/station-properties.html">
-    {% include 'transport/app/ng-templates/station-properties.html' %}
+    {% include 'transport/ng-templates/station-properties.html' %}
 </script>
 
 <script type="text/ng-template" id="munimap-transport/construction-properties.html">
-    {% include 'transport/app/ng-templates/construction-properties.html' %}
+    {% include 'transport/ng-templates/construction-properties.html' %}
 </script>
 
 <script type="text/ng-template" id="munimap-transport/layerswitcher.html">
-    {% include 'transport/app/ng-templates/layerswitcher.html' %}
+    {% include 'transport/ng-templates/layerswitcher.html' %}
 </script>
 
 <script type="text/ng-template" id="munimap/searchbox.html">
