+++ munimap/templates/munimap_transport/ng-templates/station-properties.html
