+++ munimap/templates/munimap_transport/sidebar.html
@@ -29,11 +29,11 @@
 </script>
 
 <script type="text/ng-template" id="munimap/sidebar-header.html">
-  {% include 'transport/app/ng-templates/sidebar-header.html' %}
+  {% include 'transport/ng-templates/sidebar-header.html' %}
 </script>
 
 <script type="text/ng-template" id="munimap/sidebar-footer.html">
-  {% include 'transport/app/ng-templates/sidebar-footer.html' %}
+  {% include 'transport/ng-templates/sidebar-footer.html' %}
 </script>
 
 <script type="text/ng-template" id="munimap/layerswitcher-overlays.html">
