+++ munimap/translations/messages.pot
@@ -961,12 +961,12 @@ msgid "Remove feature"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/index.html:124
-#: ../munimap_transport/munimap_transport/templates/transport/app/index.html:53
+#: ../munimap/templates/munimap_transport/transport/index.html:53
 msgid "Use ctrl + wheel to zoom"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/index.html:125
-#: ../munimap_transport/munimap_transport/templates/transport/app/index.html:54
+#: ../munimap/templates/munimap_transport/transport/index.html:54
 msgid "Use two fingers to move the map"
 msgstr ""
 
@@ -995,38 +995,38 @@ msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:117
 #: ../munimap/templates/munimap/app/ng-templates/sidebar-header.html:3
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:74
-#: ../munimap_transport/munimap_transport/templates/transport/app/ng-templates/sidebar-header.html:3
+#: ../munimap/templates/munimap_transport/transport/map.html:74
+#: ../munimap/templates/munimap_transport/transport/ng-templates/sidebar-header.html:3
 msgid "Close menu"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:117
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:74
+#: ../munimap/templates/munimap_transport/transport/map.html:74
 msgid "Open menu"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:129
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:86
+#: ../munimap/templates/munimap_transport/transport/map.html:86
 msgid "Startposition"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:250
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:148
+#: ../munimap/templates/munimap_transport/transport/map.html:148
 msgid "UTM 32"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:251
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:147
+#: ../munimap/templates/munimap_transport/transport/map.html:147
 msgid "WGS 84"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:252
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:149
+#: ../munimap/templates/munimap_transport/transport/map.html:149
 msgid "Web Mercator"
 msgstr ""
 
 #: ../munimap/templates/munimap/app/map.html:266
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:171
+#: ../munimap/templates/munimap_transport/transport/map.html:171
 msgid "Open Timetable"
 msgstr ""
 
@@ -1842,64 +1842,64 @@ msgstr ""
 msgid "From %(_from)s to %(to)s"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:134
+#: ../munimap/templates/munimap_transport/transport/map.html:134
 msgid "Restart tour"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:160
+#: ../munimap/templates/munimap_transport/transport/map.html:160
 msgid "Menu"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:160
+#: ../munimap/templates/munimap_transport/transport/map.html:160
 msgid "close"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/map.html:160
+#: ../munimap/templates/munimap_transport/transport/map.html:160
 msgid "open"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/sidebar.html:4
+#: ../munimap/templates/munimap_transport/transport/sidebar.html:4
 msgid "Overlay layers"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/sidebar.html:10
+#: ../munimap/templates/munimap_transport/transport/sidebar.html:10
 msgid "Search"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:4
+#: ../munimap/templates/munimap_transport/transport/timetable.html:4
 msgid "Timetable"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:36
+#: ../munimap/templates/munimap_transport/transport/timetable.html:36
 msgid "From"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:39
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:52
+#: ../munimap/templates/munimap_transport/transport/timetable.html:39
+#: ../munimap/templates/munimap_transport/transport/timetable.html:52
 msgid "Station"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:48
+#: ../munimap/templates/munimap_transport/transport/timetable.html:48
 msgid "To"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:60
+#: ../munimap/templates/munimap_transport/transport/timetable.html:60
 msgid "Date"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:69
+#: ../munimap/templates/munimap_transport/transport/timetable.html:69
 msgid "Time"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/timetable.html:81
+#: ../munimap/templates/munimap_transport/transport/timetable.html:81
 msgid "Search connection"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/ng-templates/station-properties.html:11
+#: ../munimap/templates/munimap_transport/transport/ng-templates/station-properties.html:11
 msgid "Select station as start"
 msgstr ""
 
-#: ../munimap_transport/munimap_transport/templates/transport/app/ng-templates/station-properties.html:18
+#: ../munimap/templates/munimap_transport/transport/ng-templates/station-properties.html:18
 msgid "Select station as stop"
 msgstr ""
