+++ munimap/views/__init__.py
@@ -8,3 +8,4 @@ from .admin import admin
 from .api import api
 from .alkis import alkis
 from .digitize import digitize
+from .transport import *
