--- munimap_transport/munimap_transport/views/__init__.py
@@ -1,1 +1,0 @@
-from .transport import *
