--- munimap_transport/pyproject.toml
@@ -1,18 +1,0 @@
-[build-system]
-requires = ['hatchling']
-build-backend = 'hatchling.build'
-
-[project]
-name = 'munimap_transport'
-version = '1.0'
-maintainers = [
-    { name = 'terrestris GmbH & Co. KG', email = 'info@terrestris.de' }
-]
-dependencies = [
-    'Flask==0.12.5',
-    'natsort==8.4.0'
-]
-classifiers = [
-    "License :: OSI Approved :: BSD License",
-    "Programming Language :: Python :: 3.9"
-]
